Change DelimitedStreamReader#readLine() to return lines with \r\n.
The default behavior was that those were being trimmed. It was hard to create
a new class responsible for this without a bigger effort so this IS a shortcut
and it changes the default behavior of HBC. On the other hand we need each
tweet JSON to end with a newline so Hadoop can pick this up.